backend/requirements.txt
#app
uvicorn==0.25.0
fastapi==0.108.0
loguru==0.7.2
python-multipart==0.0.9
fastapi-utils==0.7.0
azure-cognitiveservices-speech==1.38.0
azure-identity==1.16.1
python-dotenv==1.0.1
pytz==2024.1
typing-inspect==0.9.0
jinja2==3.1.4
emoji==2.12.1
openai==1.35.3
azure-search-documents==11.4.0
easydict==1.13
pyyaml==6.0.1
azure-data-tables==12.5.0